Add .idea to .gitignore and refactor code
The .idea directory has been added to the .gitignore file to prevent accidental commits of IDE settings. In addition, some minor updates have been made to the Python and Markdown files, including aligning table entries to the center in generate_readme.py and adding necessary spacing in the Markdown file.