Add solution for AcWing 802 problem in markdown and C++, Go languages
The new AcWing problem includes n operations that adds a number c to a position x on an infinite number line and m queries asking for the sum of all numbers in a range. The solutions in C++ and Go calculates the range sum by binary search and prefix sum after sorting and deduplicating all query and addition positions.