Add "Merge Sets" problem and solutions in C++ and Go
A new problem titled "AcWing 836. Merge Sets" has been added to the system, with the problem statement provided in markdown format. Additionally, solutions to the problem have been included in both C++ and Go languages. The problem centres around operating on two types of operations on a set - merging two elements into the same set and querying whether two elements are in the same set.