Add solution for "AcWing 837. Connected Block Number of Points"
A new markdown file has been added providing a solution for the problem "AcWing 837. Connected Block Number of Points". This problem aims to examine the understanding of disjoint set union concepts. Implementations are shown in languages C++ and Go. The markdown also contains a problem description, input/output examples, and a simple explanation of the data structures (disjoint sets) used in the solutions.